// src/statistics/dto/statistics.dto.ts
export interface MonthlySales {
  month: string;
  amount: number;
  orderCount: number;
}

export interface TopProduct {
  productId: string;
  productName: string;
  sales: number;
  productPrice: number;
}

export class StatisticsDto {
  totalUsers: number;
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  lowStockItems: number;
  recentOrders: number;
  monthlySales: MonthlySales[];
  topProducts: TopProduct[];
}